Delay and Beep streams implemented

+++ program/Core/system/base.h
@@ -12,7 +12,7 @@ extern uint8_t pointMenu;
 extern unsigned char counter;
 extern volatile uint16_t timerDelay;
 extern volatile uint16_t timerValue;
-extern volatile uint16_t beepCounter;
+extern volatile uint8_t beepCounter;
 extern uint8_t res;
 extern UINT s1;
 extern FATFS fs1;
